updeta

+++ L02/EXE4 calculator.py
@@ -17,8 +17,8 @@ elif operation == 3:
     num2 = int(input('Enter second number : '))
     print(num1,' * ',num2,' = ',num1 * num2)
 elif operation == 4:
-    num1 = int(input('Enter first number : '))
-    num2 = int(input('Enter second number : '))
+    num1 = float(input('Enter first number : '))
+    num2 = float(input('Enter second number : '))
     print(num1,' / ',num2,' = ',num1 / num2)
 else:
     print('Operation Errer!!!')
